feat(formula-input): add api route
